clickHouse test

+++ dispatcher/dataStructures.go
@@ -4,7 +4,7 @@ import "time"
 
 type MessageResult struct {
 	MessageId    string
-	ResultCode   int
+	ResultCode   uint16
 	ResponseBody []byte
 	Error        string
 	CreatedAt    time.Time
